+++ app/renderer/routes/index.js
@@ -9,7 +9,10 @@ import ChooseNetworkPage from 'renderer/pages/chooseNetworkPage';
 import SyncNodePage from 'renderer/pages/syncNodePage';
 import DevPage from 'renderer/pages/devPage';
 import HomePage from 'renderer/pages/homePage';
-import WalletRoutes from 'renderer/routes/WalletRoutes';
+import WalletLandingPage from 'renderer/pages/wallet/landingPage';
+import WalletListPage from 'renderer/pages/wallet/listPage';
+import WalletCreatePage from 'renderer/pages/wallet/create';
+import WalletConnectPage from 'renderer/pages/wallet/connect';
 import SettingsRoutes from 'renderer/routes/SettingsRoutes';
 import ROUTES from 'renderer/constants/routes';
 
@@ -34,9 +37,11 @@ class AppRoutes extends React.Component {
     return (
       <Switch>
         <Route exact path={ROUTES.ROOT} component={HomePage} />
-        <Redirect exact from={ROUTES.WALLET.ROOT} to={ROUTES.WALLET.CREATE} />
         <Redirect exact from={ROUTES.SETTINGS.ROOT} to={ROUTES.SETTINGS.GENERAL} />
-        <Route path={ROUTES.WALLET.ROOT} render={WalletRoutes} />
+        <Route exact path={ROUTES.WALLET.ROOT} component={WalletLandingPage} />
+        <Route exact path={ROUTES.WALLET.LIST} component={WalletListPage} />
+        <Route exact path={ROUTES.WALLET.CREATE} component={WalletCreatePage} />
+        <Route exact path={ROUTES.WALLET.CONNECT} component={WalletConnectPage} />
         <Route path={ROUTES.SETTINGS.ROOT} render={SettingsRoutes} />
         <Route exact path={ROUTES.TERMS_OF_USE_ACCEPTANCE} component={TosPage} />
         <Route exact path={ROUTES.CHOOSE_NETWORK} render={() => <ChooseNetworkPage />} />